refactor models to LodeRunner.Core namesapce
