+++ src/LodeRunner.Core/Models/TestRun.cs
@@ -4,7 +4,7 @@
 using System;
 using System.Collections.Generic;
 
-namespace LodeRunner.Data.Model
+namespace LodeRunner.Core.Models
 {
     /// <summary>
     /// TestRun Model.
