+++ src/LodeRunner.Core/stylecop.json
@@ -1,0 +1,20 @@
+﻿{
+  // ACTION REQUIRED: This file was automatically added to your project, but it
+  // will not take effect until additional steps are taken to enable it. See the
+  // following page for additional information:
+  //
+  // https://github.com/DotNetAnalyzers/StyleCopAnalyzers/blob/master/documentation/EnableConfiguration.md
+
+  "$schema": "https://raw.githubusercontent.com/DotNetAnalyzers/StyleCopAnalyzers/master/StyleCop.Analyzers/StyleCop.Analyzers/Settings/stylecop.schema.json",
+  "settings": {
+    "orderingRules": {
+      "usingDirectivesPlacement": "outsideNamespace"
+    },
+
+    "documentationRules": {
+      "xmlHeader": false,
+      "companyName": "Microsoft Corporation",
+      "copyrightText": "Copyright (c) Microsoft Corporation. All rights reserved.\nLicensed under the MIT License. See LICENSE in the project root for license information."
+    }
+  }
+}
