+++ src/LodeRunner.Data/Interfaces/IClientStatusService.cs
@@ -5,7 +5,7 @@ using System;
 using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
-using LodeRunner.Data.Model;
+using LodeRunner.Core.Models;
 
 namespace LodeRunner.Data.Interfaces
 {
