+++ src/LodeRunner.Data/Services/LoadTestConfigService.cs
@@ -3,8 +3,8 @@
 
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using LodeRunner.Core.Models;
 using LodeRunner.Data.Interfaces;
-using LodeRunner.Data.Model;
 
 namespace LodeRunner.Services
 {
