+++ src/LodeRunner/LodeRunner.csproj
@@ -30,7 +30,7 @@
 
   <ItemGroup>
   
-    <PackageReference Include="FluentValidation" Version="10.3.1" />
+    <PackageReference Include="FluentValidation" Version="10.3.3" />
   
     <PackageReference Include="prometheus-net.AspNetCore" Version="4.1.1" />
     <PackageReference Include="Microsoft.CodeAnalysis.NetAnalyzers" Version="5.0.3">
@@ -77,6 +77,7 @@
     <Folder Include="src\events\" />
   </ItemGroup>
   <ItemGroup>
+    <ProjectReference Include="..\LodeRunner.Core\LodeRunner.Core.csproj" />
     <ProjectReference Include="..\LodeRunner.Data\LodeRunner.Data.csproj" />
   </ItemGroup>
